crearUsuario/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
PyJWT
python-dotenv
psycopg2-binary